// src/tests/benchmarks/classes/with-optional-property.interface.ts
export interface WithOptionalProperty {
  optionalProperty1?: string;
  optionalProperty2?: string;
  optionalProperty3?: string;
  optionalProperty4?: string;
  optionalProperty5?: string;
  optionalProperty6?: string;
  optionalProperty7?: string;
  optionalProperty8?: string;
  optionalProperty9?: string;
  optionalProperty10?: string;
  optionalProperty11?: string;
  optionalProperty12?: string;
  optionalProperty13?: string;
  optionalProperty14?: string;
  optionalProperty15?: string;
  optionalProperty16?: string;
  optionalProperty17?: string;
  optionalProperty18?: string;
  optionalProperty19?: string;
  optionalProperty20?: string;
}